Added some actions.
Is declareVariable in encode.c suppose to be the same as encode_global_var_decl in his?  If so, are we doing it right?  I ask because the params are not the same.  I am thinking we can use one function for encoding both vars and functions, but in order to adapt our existing function I need to understand what is going on.